\docType{package}
\name{grealtime}
\alias{grealtime}
\alias{grealtime-package}
\title{gprediction}
\description{
  Google Analytics Real-time API client for R
}
\details{
  A Google Analytics Real-time API client for R supporting
  OAuth2.0
}
\references{
  Google Analytics Real-time API developer reference
  \url{https://developers.google.com/analytics/devguides/reporting/realtime/v3/}
}
\keyword{package}

