\name{GrtGet}
\alias{GrtGet}
\title{GrtGet
Returns real-time data for a view (profile).}
\usage{
  GrtGet(oauth, ids, metrics = "ga:activeVisitors",
    dimensions = NULL, filters = NULL, max_results = NULL,
    sort_by = NULL)
}
\description{
  GrtGet Returns real-time data for a view (profile).
}

